fix(docs-site): move buildLlmsTxt out of the route module
Next.js route modules may only export request handlers + reserved config —
exporting the shared `buildLlmsTxt()` helper from app/api/llms.txt/route.ts
(added in #2329 for the /api/llms.txt.md twin) fails the route-type check:
"Route app/api/llms.txt/route.ts does not match the required types of a
Next.js Route." This reddened the Docs Site Build on main (non-required check,
so it slipped past auto-merge on #2329/#2331/#2333/#2334) and blocks the
release-please PR #2332.

Move buildLlmsTxt to docs/site/lib/llms-txt.ts; both routes import it from
there. A fresh `next build` (no stale .next cache — which is why the earlier
--webpack build missed it) now compiles /api/llms.txt + /api/llms.txt.md clean.

Verified: next build exit 0, "Compiled successfully", both routes built.

Refs #2329
