+++ docs/site/app/api/llms.txt.md/route.ts
@@ -1,6 +1,6 @@
 // Generated by OrchestKit Claude Plugin
 // Created: 2026-06-08
-import { buildLlmsTxt } from "@/app/api/llms.txt/route";
+import { buildLlmsTxt } from "@/lib/llms-txt";
 
 // /api/llms.txt.md — Markdown twin of /api/llms.txt. Same body (heading-led,
 // starts with a top-level `#`), but served as text/markdown so agents and
