+++ docs/site/lib/llms-txt.ts
@@ -1,0 +1,45 @@
+// Generated by OrchestKit Claude Plugin
+// Created: 2026-06-10
+
+import { SITE } from "@/lib/constants";
+
+// Shared generator for the API-surface llms.txt. Lives in lib (not a route file)
+// so both /api/llms.txt and its Markdown twin /api/llms.txt.md can import it —
+// Next.js route modules may only export request handlers + reserved config, so a
+// shared helper exported from a route.ts breaks the route-type check at build.
+export function buildLlmsTxt(): string {
+	const d = SITE.domain;
+	return [
+		`# ${SITE.name} API`,
+		"",
+		`> Public, read-only HTTP API over the ${SITE.name} documentation. No authentication. Errors use the RFC 9457 Problem Details shape, served as \`application/json\`. Current version: v1.`,
+		"",
+		"## Endpoints",
+		"",
+		`- \`GET ${d}/api/search?query=...&limit=10\` — full-text docs search (JSON array; \`X-Total-Count\` header; \`limit\` paginates). Versioned alias: \`/api/v1/search\`.`,
+		`- \`POST ${d}/ask\` — NLWeb natural-language query → JSON with \`_meta\`; send \`Accept: text/event-stream\` for SSE streaming.`,
+		`- \`POST ${d}/api/mcp\` — MCP server (Streamable HTTP / JSON-RPC): \`initialize\`, \`tools/list\`, \`tools/call\`. Tools are read-only (\`readOnlyHint: true\`).`,
+		`- \`GET ${d}/api/openapi\` — OpenAPI 3.1 description of the API.`,
+		`- \`GET ${d}/api/health\` — liveness check.`,
+		"",
+		"## Use cases",
+		"",
+		"- Answer a user's natural-language question about OrchestKit (`/ask`).",
+		"- Find the right documentation page before fetching its Markdown (`/api/search` → append `.md` to the page URL).",
+		"- Connect an MCP-capable agent (Claude, etc.) to search docs and read pages natively (`/api/mcp`).",
+		"",
+		"## Constraints",
+		"",
+		"- Read-only: there are no write, payment, or account endpoints.",
+		"- No authentication and no per-key rate limit; the API is served behind a CDN. Be a good citizen: cache responses (most carry `Cache-Control`) and back off on transient `5xx` with exponential retry (e.g. 1s, 2s, 4s).",
+		"- Search is over documentation content only (not source code).",
+		"",
+		"## Discovery",
+		"",
+		`- API catalog (RFC 9727): ${d}/.well-known/api-catalog`,
+		`- MCP server card: ${d}/.well-known/mcp/server-card.json`,
+		`- A2A agent card: ${d}/.well-known/agent-card.json`,
+		`- Auth (none): ${d}/auth.md`,
+		"",
+	].join("\n");
+}
